major cleanup of BKPianoSamplerVoice::startNote, to make it more readable and to fix clicks of various sorts

+++ bk_JUCE/bitKlavier/Source/BKPianoSampler.h
@@ -118,8 +118,8 @@ public:
                     BKNoteType bktype,
                     uint64 startingPosition,
                     uint64 length,
-                    int voiceRampOn,
-                    int voiceRampOff,
+                    uint64 voiceRampOn,
+                    uint64 voiceRampOff,
                     BKSynthesiserSound* sound
                     ) override;
     
@@ -131,10 +131,10 @@ public:
                     BKNoteType bktype,
                     uint64 startingPosition,
                     uint64 length,
-                    int adsrAttack,
-                    int adsrDecay,
+                    uint64 adsrAttack,
+                    uint64 adsrDecay,
                     float adsrSustain,
-                    int adsrRelease,
+                    uint64 adsrRelease,
                     BKSynthesiserSound* sound
                     ) override;
